Feat: Internalize removeClientFunctionCallId into content processor in adk-js
